third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-automatic-minimum-for-auto-rows-expected.txt



Check that min-height is honored when sizing auto rows.
PASS window.getComputedStyle(gridAuto, '').getPropertyValue('grid-template-rows') is "40px"
PASS window.getComputedStyle(gridAutoItemSmallerThanMinSize, '').getPropertyValue('grid-template-rows') is "40px"
PASS window.getComputedStyle(gridMinMaxAutoAuto, '').getPropertyValue('grid-template-rows') is "40px"
PASS window.getComputedStyle(gridMinMaxAutoMaxContent, '').getPropertyValue('grid-template-rows') is "30px"
PASS window.getComputedStyle(gridMinMaxMinContentAuto, '').getPropertyValue('grid-template-rows') is "60px"
PASS window.getComputedStyle(gridAutoMultipleItems, '').getPropertyValue('grid-template-rows') is "60px"
PASS window.getComputedStyle(gridMinMaxAutoAutoMultipleItemsOneWithoutMinHeight, '').getPropertyValue('grid-template-rows') is "50px"
PASS window.getComputedStyle(gridMinMaxAutoMaxContentMultipleItemsOneWithAutoMinHeight, '').getPropertyValue('grid-template-rows') is "60px"
PASS window.getComputedStyle(gridAutoAndAutoFixedHeightChildren, '').getPropertyValue('grid-template-rows') is "200px 50px"


Check that min-height is honored when sizing auto rows and spanning grid items.
PASS window.getComputedStyle(gridAutoAndAutoOneSpanningOneNonSpannig, '').getPropertyValue('grid-template-rows') is "10px 40px"
PASS window.getComputedStyle(gridAutoAndAutoOneSpanningMultipleNonSpanning, '').getPropertyValue('grid-template-rows') is "40px 20px"
PASS window.getComputedStyle(gridAutoAndMinMaxAutoAutoOneSpanningOneNonSpanning, '').getPropertyValue('grid-template-rows') is "40px 30px"
PASS window.getComputedStyle(gridAutoAndMinMaxAutoAutoMultipleSpanning, '').getPropertyValue('grid-template-rows') is "35px 35px"
PASS window.getComputedStyle(gridMinMaxAutoMaxContentAndAutoOneSpanningMultipleNonSpanning, '').getPropertyValue('grid-template-rows') is "25px 65px"
PASS window.getComputedStyle(gridMinMaxAutoMaxContentAndAutoMultipleSpanningMultipleNonSpanning, '').getPropertyValue('grid-template-rows') is "50px 40px"


Check the interactions between height and min-height and auto tracks.
PASS window.getComputedStyle(gridAutoWithFixedHeightChild, '').getPropertyValue('grid-template-rows') is "60px"
PASS window.getComputedStyle(gridAutoWithFixedHeightChildAndMinHeight, '').getPropertyValue('grid-template-rows') is "60px"
PASS window.getComputedStyle(gridAutoWithFixedHeightChildAndHigherMinHeight, '').getPropertyValue('grid-template-rows') is "90px"
PASS window.getComputedStyle(gridAutoAndAutoOneSpanningFixedHeight, '').getPropertyValue('grid-template-rows') is "25px 25px"
PASS window.getComputedStyle(gridAutoAndAutoOneSpanningFixedHeightAndMinHeight, '').getPropertyValue('grid-template-rows') is "30px 30px"
PASS window.getComputedStyle(gridAutoAndAutoOneSpanningFixedHeightAndHigherMinHeight, '').getPropertyValue('grid-template-rows') is "35px 35px"


Check that borders and paddings are considering when computing min sizes.
PASS window.getComputedStyle(gridAutoFixedMinHeightWithBorder, '').getPropertyValue('grid-template-rows') is "50px"
PASS window.getComputedStyle(gridAutoFixedMinHeightWithPadding, '').getPropertyValue('grid-template-rows') is "56px"
PASS window.getComputedStyle(gridAutoFixedMinHeightWithBorderAndPadding, '').getPropertyValue('grid-template-rows') is "66px"
PASS window.getComputedStyle(gridAutoAutoMinHeightWithBorder, '').getPropertyValue('grid-template-rows') is "40px"
PASS window.getComputedStyle(gridAutoAutoMinHeightWithPadding, '').getPropertyValue('grid-template-rows') is "46px"
PASS window.getComputedStyle(gridAutoAutoMinHeightWithBorderAndPadding, '').getPropertyValue('grid-template-rows') is "56px"
PASS window.getComputedStyle(gridAutoMaxContentMinHeightWithBorder, '').getPropertyValue('grid-template-rows') is "30px"
PASS window.getComputedStyle(gridAutoMaxContentMinHeightWithPadding, '').getPropertyValue('grid-template-rows') is "36px"
PASS window.getComputedStyle(gridAutoMaxContentMinHeightWithBorderAndPadding, '').getPropertyValue('grid-template-rows') is "46px"
PASS successfullyParsed is true

TEST COMPLETE

XX
XXX
XX
XXX
XXXX
XX
XX
XX
X
X
X
XXX
X
XX
XX
X
X
XXXX
X
XX
XXX
X
X
XX
XXX
XX
XXX
XX
XX XX
XXXXXX X XXX
XX XX XX
X X X X
XX XX
X XXX XX
XXX XXXX
XX XX XX
XXXXX X XXXXX
X XXX XX
XXXXX XXXXX
XX XX
XX XX XX XX
XXX
X XX XXX
X XX X
XXX X
XXX X
XXX X
XX XXX XX XXX XX XXX
XX XXX XX
XX XXX XX X XX
XXXXXX
XXXXXX
XXXXXX
XX
XXX X
XX
XXX X
XX
XXX X
XXX X
XXX X
XXX X
